+++ CMP2801M-Week5/Triangle.h
@@ -4,5 +4,5 @@ class Triangle : public Shape
 {
 	public:
 		Triangle(float width, float height);
-		double getArea();
+		float getArea();
 };
